main logic switch to sdk - cursor
