+++ airflow-core/src/airflow/example_dags/example_anomaly_detection.py
@@ -26,8 +26,7 @@ import datetime
 import time
 from pathlib import Path
 
-from airflow.anomaly_detection.anomalydetector import AnomalyDetector, ThresholdAnomaly
-from airflow.sdk import DAG, Param, TriggerRule, task
+from airflow.sdk import DAG, AnomalyDetector, Param, ThresholdAnomaly, TriggerRule, task
 
 # [START params_trigger]
 with DAG(
