+++ airflow-core/src/airflow/models/task_instance_anomaly.py
@@ -18,9 +18,13 @@
 from __future__ import annotations
 
 from datetime import datetime
+from typing import TYPE_CHECKING, Any
 
-from sqlalchemy import Boolean, Integer, String, UniqueConstraint
+from sqlalchemy import Boolean, Integer, String, UniqueConstraint, select
 from sqlalchemy.orm import Mapped, mapped_column
+
+if TYPE_CHECKING:
+    from sqlalchemy.orm import Session
 
 from airflow._shared.timezones import timezone
 from airflow.models.base import Base
@@ -65,3 +69,42 @@ class TaskInstanceAnomaly(Base):
             name="uq_task_instance_anomaly_dag_run_task_map_try",
         ),
     )
+
+
+def record_task_instance_anomaly(
+    *,
+    session: Session,
+    task_instance: Any,
+    is_anomalous: bool,
+    detector_name: str,
+    reason: str | None,
+) -> None:
+    """Insert or update the TaskInstanceAnomaly row for the given TaskInstance."""
+    identity = {
+        "dag_id": task_instance.dag_id,
+        "run_id": task_instance.run_id,
+        "task_id": task_instance.task_id,
+        "map_index": getattr(task_instance, "map_index", -1),
+        "try_number": getattr(task_instance, "try_number", 0),
+    }
+    anomaly_values = {
+        "is_anomalous": is_anomalous,
+        "detector_name": detector_name,
+        "reason": reason,
+    }
+
+    row = session.scalar(
+        select(TaskInstanceAnomaly).where(
+            TaskInstanceAnomaly.dag_id == identity["dag_id"],
+            TaskInstanceAnomaly.run_id == identity["run_id"],
+            TaskInstanceAnomaly.task_id == identity["task_id"],
+            TaskInstanceAnomaly.map_index == identity["map_index"],
+            TaskInstanceAnomaly.try_number == identity["try_number"],
+        )
+    )
+
+    if row is None:
+        session.add(TaskInstanceAnomaly(**identity, **anomaly_values))
+    else:
+        for field, value in anomaly_values.items():
+            setattr(row, field, value)
